Oops! Removed mac compile warning.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@2635 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/PlayerEntity.m
@@ -376,7 +376,7 @@ static PlayerEntity *sSharedPlayer = nil;
 	
 	NSMutableArray* missileRoles = [NSMutableArray arrayWithCapacity:max_missiles];
 	
-	for (i = 0; i < max_missiles; i++)
+	for (i = 0; i < (int)max_missiles; i++)
 	{
 		if (missile_entity[i])
 		{
